+++ lagom/interfaces.py
@@ -198,4 +198,5 @@ TypeResolver = Union[
     SpecialDepDefinition[T],  # From the definitions module
     SpecialDepDefinition[NoReturn],  # Some types don't resolve
     T,  # Just an instance of the type - A singleton
+    List[Callable],  # A collection of functions - maps to FunctionCollection[F]
 ]